requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.74.1
h11==0.13.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
pydantic==1.9.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
